+++ call.go
@@ -100,6 +100,11 @@ type Call encoding.Call
 // context.
 //
 // The object is valid only as long as the request is ongoing.
+//
+// Testing
+//
+// To test functions which use CallFromContext, use yarpctest.ContextWithCall
+// to build contexts compatible with this function.
 func CallFromContext(ctx context.Context) *Call {
 	return (*Call)(encoding.CallFromContext(ctx))
 }