+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - gRPC inbound supports introspection, suitable for debug pages.
+- yarpctest: Add `ContextWithCall` function to ease testing of functions that
+  use `yarpc.CallFromContext`.
 
 ## [1.44.0] - 2020-02-27
 ### Added